test: 닉네임 중복 검사 Service Test 작성
