+++ src/test/java/com/hongik/service/user/UserServiceTest.java
@@ -1,0 +1,106 @@
+package com.hongik.service.user;
+
+import com.hongik.domain.user.Role;
+import com.hongik.domain.user.User;
+import com.hongik.domain.user.UserRepository;
+import com.hongik.dto.user.request.NicknameRequest;
+import com.hongik.dto.user.request.UserCreateRequest;
+import com.hongik.dto.user.response.UserResponse;
+import com.hongik.exception.AppException;
+import com.hongik.service.UserService;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+import static org.assertj.core.api.Assertions.*;
+
+@SpringBootTest
+class UserServiceTest {
+
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    @AfterEach
+    void tearDown() {
+        userRepository.deleteAllInBatch();
+    }
+
+    @DisplayName("회원가입을 완료한다.")
+    @Test
+    void createUser() {
+        // given
+        UserCreateRequest request = UserCreateRequest.builder()
+                .username("test@email.com")
+                .password("password")
+                .nickname("nickname")
+                .department("department")
+                .build();
+
+        // when
+        UserResponse userResponse = userService.signUp(request);
+
+        // then
+        assertThat(userResponse.getId()).isNotNull();
+        assertThat(userResponse)
+                .extracting("username", "nickname", "department")
+                .containsExactlyInAnyOrder("test@email.com", "nickname", "department");
+    }
+
+    @DisplayName("회원가입을 진행할 때 중복된 닉네임이 존재하면 예외가 발생한다.")
+    @Test
+    void createUserWithDuplicationNickname() {
+        // given
+        final String nickname = "nickname";
+        User user = createUser("test@email.com", "password", nickname);
+        userRepository.save(user);
+
+        UserCreateRequest request = UserCreateRequest.builder()
+                .username("test@email.com")
+                .password("password")
+                .nickname(nickname)
+                .department("department")
+                .build();
+
+        // when // then
+        assertThatThrownBy(() -> userService.signUp(request))
+                .isInstanceOf(AppException.class)
+                .hasMessage("이미 존재하는 닉네임입니다.");
+    }
+
+    @DisplayName("닉네임 중복검사를 할 때 이미 닉네임이 존재하면 예외가 발생한다.")
+    @Test
+    void checkNicknameDuplication() {
+        // given
+        final String nickname = "nickname";
+        User user = createUser("test@email.com", "password", nickname);
+        userRepository.save(user);
+
+        NicknameRequest request = NicknameRequest.builder()
+                .nickname(nickname)
+                .build();
+
+        // when // then
+        assertThatThrownBy(() -> userService.checkNicknameDuplication(request.getNickname()))
+                .isInstanceOf(AppException.class)
+                .hasMessage("이미 존재하는 닉네임입니다.");
+    }
+
+    private User createUser(final String username, final String password, final String nickname) {
+        return User.builder()
+                .username(username)
+                .password(password)
+                .role(Role.USER)
+                .department("department")
+                .nickname(nickname)
+                .build();
+    }
+}
